[PATCH] x86-64 merge
This patch depends on the i386 MTRR driver cleanup I sent earlier.

 - Support non executable mappings for x86-64. data/heap are non executable
   by default now.
 - Beginnings of software suspend from Pavel (not working yet)
 - Support generic compat functions and remove some shared code
   in the 32bit emulation (Stephen Rothwell)
 - Support hugetlbfs
 - Some makefile updates
 - Make sure all 32bit emulation functions return long, not int.
   This fixes some problems with ERESTARTNOSYS.et.al. leaking to userspace.
 - Add new system calls.
 - Fix long standing fs/gs context switch bugs (thanks to Karsten Keil
   for helping to fix that mess). Also make sure the gs selector is
   set to 0 after an exec.
 - Simplify TLS switching
 - Paranoid CPUID check at bootup
 - Reorder scatterlist to be more space efficient (Jes Soerensen)
 - Enlarge 32bit address space to full 4GB.
 - Beginnings of 32bit SYSCALL support (not completely working yet
   and vsyscall page miss yet)
 - Various merges from i386
 - New module loader
 - Support threaded core dump (XMM saving for 32bit programs doesn't
   work, but it appears to be broken on i386 too)
 - Fix bug in signal stack rounding
 - Remove DRM 32bit emulation.
 - Use MTRR driver from i386
 - Use bootflag.c from i386
 - Various other fixes and cleanups.

+++ arch/x86_64/ia32/ia32_signal.c
@@ -76,7 +76,7 @@ static int ia32_copy_siginfo_to_user(siginfo_t32 *to, siginfo_t *from)
 	}
 }
 
-asmlinkage int
+asmlinkage long
 sys32_sigsuspend(int history0, int history1, old_sigset_t mask, struct pt_regs regs)
 {
 	sigset_t saveset;
@@ -97,7 +97,7 @@ sys32_sigsuspend(int history0, int history1, old_sigset_t mask, struct pt_regs r
 	}
 }
 
-asmlinkage int
+asmlinkage long
 sys32_sigaltstack(const stack_ia32_t *uss_ptr, stack_ia32_t *uoss_ptr, 
 				  struct pt_regs regs)
 {
@@ -227,7 +227,7 @@ badframe:
 	return 1;
 }
 
-asmlinkage int sys32_sigreturn(struct pt_regs regs)
+asmlinkage long sys32_sigreturn(struct pt_regs regs)
 {
 	struct sigframe *frame = (struct sigframe *)(regs.rsp - 8);
 	sigset_t set;
@@ -256,7 +256,7 @@ badframe:
 	return 0;
 }	
 
-asmlinkage int sys32_rt_sigreturn(struct pt_regs regs)
+asmlinkage long sys32_rt_sigreturn(struct pt_regs regs)
 {
 	struct rt_sigframe *frame = (struct rt_sigframe *)(regs.rsp - 4);
 	sigset_t set;
